IGNITE-11512 Add counter left partition for index rebuild in CacheGroupMetricsMXBean. Delete instanceof, add CacheGroupMetrics
