+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CacheGroupContext.java
@@ -186,6 +186,9 @@ public class CacheGroupContext {
     /** */
     private volatile boolean hasAtomicCaches;
 
+    /** Store cache group metrics. */
+    private final CacheGroupMetricsImpl metrics;
+
     /**
      * @param ctx Context.
      * @param grpId Group ID.
@@ -246,6 +249,8 @@ public class CacheGroupContext {
         mvccEnabled = ccfg.getAtomicityMode() == TRANSACTIONAL_SNAPSHOT;
 
         log = ctx.kernalContext().log(getClass());
+
+        metrics=new CacheGroupMetricsImpl();
 
         mxBean = new CacheGroupMetricsMXBeanImpl(this);
 
@@ -1290,5 +1295,12 @@ public class CacheGroupContext {
      */
     public boolean hasAtomicCaches() {
         return hasAtomicCaches;
+    }
+
+    /**
+     * @return Metrics.
+     */
+    public CacheGroupMetricsImpl metrics0() {
+        return metrics;
     }
 }
