+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CacheGroupMetricsMXBeanImpl.java
@@ -25,7 +25,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.LongAdder;
 import org.apache.ignite.cache.CacheMode;
 import org.apache.ignite.cluster.ClusterNode;
@@ -51,9 +50,6 @@ public class CacheGroupMetricsMXBeanImpl implements CacheGroupMetricsMXBean {
 
     /** */
     private final GroupAllocationTracker groupPageAllocationTracker;
-
-    /** Number of partitions need processed for finished indexes create or rebuilding. */
-    private final AtomicLong indexBuildCountPartitionsLeft;
 
     /** Interface describing a predicate of two integers. */
     private interface IntBiPredicate {
@@ -116,8 +112,6 @@ public class CacheGroupMetricsMXBeanImpl implements CacheGroupMetricsMXBean {
         }
         else
             this.groupPageAllocationTracker = new GroupAllocationTracker(AllocatedPageTracker.NO_OP);
-
-        indexBuildCountPartitionsLeft=new AtomicLong();
     }
 
     /** {@inheritDoc} */
@@ -387,20 +381,7 @@ public class CacheGroupMetricsMXBeanImpl implements CacheGroupMetricsMXBean {
 
     /** {@inheritDoc} */
     @Override public long getIndexBuildCountPartitionsLeft() {
-        return indexBuildCountPartitionsLeft.get();
-    }
-
-    /** Set number of partitions need processed for finished indexes create or rebuilding. */
-    public void setIndexBuildCountPartitionsLeft(long indexBuildCountPartitionsLeft) {
-        this.indexBuildCountPartitionsLeft.set(indexBuildCountPartitionsLeft);
-    }
-
-    /**
-     * Commit the complete index building for partition.
-     * @return Decrement number of partitions need processed for finished indexes create or rebuilding.
-     */
-    public long decIndexBuildCountPartitionsLeft() {
-        return indexBuildCountPartitionsLeft.decrementAndGet();
+        return ctx.metrics0().getIndexBuildCountPartitionsLeft();
     }
 
     /**
