+++ modules/core/src/main/java/org/apache/ignite/mxbean/CacheGroupMetricsMXBean.java
@@ -21,12 +21,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.apache.ignite.internal.processors.cache.CacheGroupContext;
+import org.apache.ignite.internal.processors.cache.CacheGroupMetrics;
 
 /**
  * This interface defines JMX view on {@link CacheGroupContext}.
  */
 @MXBeanDescription("MBean that provides access to cache group descriptor.")
-public interface CacheGroupMetricsMXBean {
+public interface CacheGroupMetricsMXBean extends CacheGroupMetrics {
     /**
      * Gets cache group id.
      *
@@ -191,12 +192,4 @@ public interface CacheGroupMetricsMXBean {
      */
     @MXBeanDescription("Storage space allocated for group adjusted for possible sparsity, in bytes.")
     public long getSparseStorageSize();
-
-    /**
-     * @return Number of partitions need processed for finished indexes create or rebuilding.
-     * It is calculated as the number of local partition minus the processed.
-     * A value of 0 indicates that the index is built.
-     */
-    @MXBeanDescription("Number of partitions need processed for finished indexes create or rebuilding.")
-    public long getIndexBuildCountPartitionsLeft();
 }
